backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36,<3.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic>=2.11.7,<3.0
pydantic-settings>=2.3.0
email-validator==2.1.0
python-multipart==0.0.6
Pillow>=10.0.0
#alembic==1.13.1
